Reorganise pages and kill copy file
