+++ identity/webapp/views/pages/validated/index.tsx
@@ -1,6 +1,8 @@
 import { NextPage } from 'next';
+import { FunctionComponent } from 'react';
 
 import { useUserContext } from '@weco/common/contexts/UserContext';
+import { prismicPageIds } from '@weco/common/data/hardcoded-ids';
 import Button from '@weco/common/views/components/Buttons';
 import {
   ContaineredLayout,
@@ -8,11 +10,8 @@ import {
 } from '@weco/common/views/components/Layout';
 import Space from '@weco/common/views/components/styled/Space';
 import {
-  ValidatedFailedText,
-  ValidatedSuccessText,
-} from '@weco/identity/utils/copy';
-import {
   Container,
+  SectionHeading,
   Wrapper,
 } from '@weco/identity/views/components/styled/Layouts';
 import IdentityPageLayout from '@weco/identity/views/layouts/IdentityPageLayout';
@@ -22,6 +21,19 @@ export type Props = {
   message: string | string[];
   isNewSignUp: boolean;
 };
+
+const ValidatedFailedText: FunctionComponent<{
+  message: string | string[];
+}> = ({ message }) => (
+  <>
+    <SectionHeading as="h1">Failed to verify email</SectionHeading>
+    <p>{message}</p>
+    <p>
+      If you need help, please{' '}
+      <a href="mailto:library@wellcomecollection.org">contact the library</a>.
+    </p>
+  </>
+);
 
 const ValidatedPage: NextPage<Props> = ({ success, message, isNewSignUp }) => {
   const { state: userState } = useUserContext();
@@ -38,7 +50,33 @@ const ValidatedPage: NextPage<Props> = ({ success, message, isNewSignUp }) => {
             <Wrapper>
               {success || urlUsed ? (
                 <>
-                  <ValidatedSuccessText isNewSignUp={isNewSignUp} />
+                  <>
+                    <SectionHeading as="h1">Email verified</SectionHeading>
+                    <p>Thank you for verifying your email address.</p>
+
+                    {isNewSignUp && (
+                      <div data-testid="new-sign-up">
+                        <p>
+                          You can now request up to 15 items from our closed
+                          stores in the library.
+                        </p>
+                        <p>
+                          To complete your membership and access subscription
+                          databases, e-journals and e-books, you’ll need to
+                          bring a form of photo identification (ID) and proof of
+                          your address to our admissions desk when you visit.
+                          The identification we accept is detailed on our{' '}
+                          <a
+                            href={`https://wellcomecollection.org/collections/${prismicPageIds.register}`}
+                          >
+                            Library membership page
+                          </a>
+                          .
+                        </p>
+                      </div>
+                    )}
+                  </>
+
                   <Button
                     variant="ButtonSolidLink"
                     link="/account"
